// api/consolidated/models/EventAnalytics.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type EventAnalytics = {
    eventName?: string | null;
    count?: number;
    uniqueUsers?: number;
    average?: number;
    lastOccurrence?: string;
    trend?: string | null;
    metrics?: Record<string, any> | null;
};